requirements.txt
numpy
pillow
python-dotenv
pyvips
requests